Refactor EEType
Split EEType into two pieces: the non-constructed EETypeNode and the
constructed ConstructedEETypeNode.
Generic type definitions get their own GenericDefinitionEETypeNode node
which matches the the one produced by the binder.

Both the constructed and the generic nodes inherit from EETypeNode.
GenericDefinitionEETypeNode could exist without inheriting from
EETypeNode, however this simplifies code that expects to get back a
common usable ObjectNode/ISymbolNode from the NodeFactory. For example,
TypeMetadataMapNode.GetData would require special casing for the
different EEType node kinds if the various EEType nodes were independent
of eachother.
