+++ src/ILCompiler.Compiler/src/Compiler/DependencyAnalysis/NodeFactory.cs
@@ -94,7 +94,14 @@ namespace ILCompiler.DependencyAnalysis
             {
                 if (_compilationModuleGroup.ContainsType(type))
                 {
-                    return new EETypeNode(type, false);
+                    if (type.IsGenericDefinition)
+                    {
+                        return new GenericDefinitionEETypeNode(type);
+                    }
+                    else
+                    {
+                        return new EETypeNode(type);
+                    }
                 }
                 else
                 {
@@ -106,7 +113,7 @@ namespace ILCompiler.DependencyAnalysis
             {
                 if (_compilationModuleGroup.ContainsType(type))
                 {
-                    return new EETypeNode(type, true);
+                    return new ConstructedEETypeNode(type);
                 }
                 else
                 {
